Feature: Add `size` formatter
The `size` formatter returns the length of an array or string, or the
number of keys in an object.

Some APIs do not return a summary of their contents but only the full
list of contents. The `size` formatter can be used to still show such
count on a widget.