Update opnsense.md to reflect effective privileges section
Amend notes on effective privileges that need to be selected for OPNSense API. It appears that in the 24.7.x release the "Status: Traffic Graph" privilege no longer exists. However, this appears to have been replaced / is covered by the Reporting: Traffic privilege grouping, which covers the /api/diagnostics/traffic/* endpoint(s).