Moonraker - added print time, filename, and disabled showing layer data when none is available
Added print time block that is available from print_stats endpoint.
Added a check whether current and total layer info is available. It is
not returned by the api if it is not explicitly configured in the
slicer. Most of the time that is not a default setting.
Added a filename display that shows the current file that is being
printed. Similar layout as in jellyfin widget.