More precise information in kubernetes.md
More precise information in kubernetes.md about the configuration. The current information about ingress being optional is misleading. It suggests that you won't have to set ingress, because it is the default. This is not true. If not set, no annotation will be evaluated and the elements won't appear on the homepage.