Added functionality to hdhomerun
Added the ability to see what each tuner is tuned to.  To accomplish this I added four mapping elements.  tunerone, tunertwo, tunerthree, tunerfour.  The element will return the channel number and name (e.g. 5.1 - WABC) or frequency (503 MHz).

I updated the language files so that the element returns #1, #2, #3 or $4 for each tuner .